Added Multilanguage resource
This resource adds logic to make the application multi-language aware.
If the main domain is accessed, the application will make a guess based
on the browser's language setting, and redirect to the correct subdomain
for the language based on ISO639 language codes. It also distinguishes
between Traditional and Simplified Chinese scripts.

This resource is only used when it has been configured in the
application's config.ini file.