Enabling PHPUnit and Phing
Started using PHPUnit to test the library. There is a phpunit.xml for
normal phpunit testing on the console, but I have added a build.xml file
for the Phing build engine. It contains a testing task, that runs the
PHPUnit testing framework, and generates a test result report, and a
code coverage report, both in HTML.