Multi-language aware partial view helper
This partial view helper will check if the multi-language resource is
being used, and try to find a translated version of the given partial,
by inserting a language code in the filename of the partial. If the
given filename was partial.phtml, it will try to find and use
partial-<lang>.phtml.