Introduced translated data members
With the $translations static variable, the programmer can tell the
model which entries are multi-lingual, i.e. has a value for every
supported language. The magic setter and getter methods check this
variable to see how to treat the data member and what to set or return.

Also added a convenience function to get the language code for
applications that do not use the MultiLanguage resource. The fallback
language is the locale's language.