Added the DebugFilter resource
The DebugFilter resource checks whether the ZFE_Core::dump() function is
allowed to output debug messages.

The ZFE_Util_Core class has been renamed to ZFE_Core for brevity, and
files that use this class, as well as test units that use this class,
have been updated to reflect the change.