Updated the Mongo resource and model
Introduced the concept of an identifier, and based on this cache loaded
entities into a process cache, so that a subsequent call within the same
process will quickly return this cached entry.

A save() call will remove the entry from the cache, so that a subsequent
call will fetch the updated data from the database again.

Added support to filter a given field on multiple values, using the $in
modifier.