Added a time() function
This registers timings in a $cumulative array, and in the dump,
calculates the duration between two subsequent time() calls.

This needs a unique ID per call, so would not work in loops.
TODO: Make it work so that it can be used in loops easily without having
to supply a separate ID.