Module-specific bootstrap code
The default Zend Framework architecture runs all the modules' bootstrap
code during the project's initialization, even when the router has
determined that the request is going to be executed by a certain module.
This is based on Chris Woodford's article on module bootstrapping:

http://offshootinc.com/blog/2011/02/11/modul-bootstrapping-in-zend-framework/

Basically, you need a Bootstrap.php for your modules, with the classname
<module>_Bootstrap, in application/modules/<module>/Bootstrap.php. This
class extends ZFE_Module_Bootstrap. Module-specific code that is going
to be used on all modules, should be done here.

Module-specific code that is going to be used inside the module only,
should be put in <module>_Bootstrap_Initializer. This goes into
application/modules/<module>/bootstrap/Initializer.php, and extends the
ZFE_Module_Initializer class.

The ZFE_Plugin_ActiveModule plugin will automatically be loaded by the
library's Bootstrap, so if you have a ZFE_Bootstrap::run() line in your
main application's bootstrap (application/Bootstrap.php) file, then it
will be run.