Improving bootstrap process
Module configuration was not working totally as expected. I'll leave the
separate bootstrapping process in the code, but the module.ini loading
has to be done another way.

One solution is to use a different public/ directory for each module,
and in each of those directories initialize the Zend_Application with an
array of the common application ini file, and the module-specific ini
file. The virtual host configuration needs to be updated too to use a
different module path, based on the hostname. At least, in my example.