Added a special tracking for updates
When fields are updated, they are added to a tracking variable. After
saving the object, this tracking variable is processed and the model
checks if there is an on*Updated function for these variables, and
executes them. This is useful for having to do something special after
something has changed.

It is a bit better than creating special _set* functions as it only
executes them when the data is actually saved into the database.