Option to get translated language names or not
Improved the getLanguages() function to return a list of languages with
their names - whether the language names should be translated into their
respective languages or not. This case is good for the language switch
pulldown, but the other case (with language names not translated) is
good for language filter pulldowns.

Also made it so that the plugin is loaded at a later stage, to give
other plugins (such as the authentication plugin) a chance to determine
the logged-in user - and therefore the user's language setting.